added DTO for account object
